// generated/api/admin/models/dynamicRow.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * BDUI API
 * Спецификация OpenAPI для Backend-Driven UI  приложения

 * OpenAPI spec version: 1.0.0
 */
import type { DynamicCompositeBase } from './dynamicCompositeBase';
import type { DynamicRowAllOf } from './dynamicRowAllOf';
import type { DynamicRowType } from './dynamicRowType';

export type DynamicRow = DynamicCompositeBase &
  DynamicRowAllOf & {
    type: DynamicRowType;
  };